Remove 'http:' in `src` attr of `img` tag.
Imgur removed the 'http:' in the `src` attributes of the `img` tags in
albums so it's now "protocol relative", and use the protocol used in
the request.  This broke `imguralbum.py`, though.

This commit removes the 'http:' in the `findall` matcher of album
images and adds it when downloading the images, and makes
`imguralbum.py` work again.